exchanged [subId] and [id] files > new vs old sub page code
